Fix bugs found in backend audit: (1) RouteMatrix.Serialize dropped RoutePrice, so server_backend read every route as price 0 and lowest-price route selection silently no-oped -- serialize it in route matrix version 5 and randomize it in GenerateRandomRouteMatrix so the round-trip test guards it. (2) SDK_SessionData.Serialize dropped PrevPacketsOutOfOrder*, so per-slice out of order deltas were computed against zero and RealOutOfOrder reported cumulative counts (could exceed 100%) -- serialize them in session data version 8. (3) Unguarded *debug deref in GetRandomBestRoute_LowestPrice panicked for non-debug buyers when all candidate routes were price >= 255 (latent until fix 1). (4) LongSessionUpdate was set by a defer that ran AFTER SessionUpdate_Post sent the analytics messages, so the flag was always false in bigquery -- moved the timing check into the post defer, before Post. (5) session_cruncher called RedisCountersWatcher getters without Lock() while the watcher thread swaps the underlying maps (race + potential index out of range when buyer keys grow). (6) leaderInstanceId was written/read without the leader mutex. (7) Session data RouteNumRelays was bounded by SDK_MaxTokens (7) but RouteRelayIds only has SDK_MaxRelaysPerRoute (5) entries -- same bit width, so no wire change. (8) Duplicate SDK_HandlerEvent value 11 (DatacenterNotEnabled vs CouldNotReadServerInitRequestPacket) -- renumbered. Also: NextRouteRelays in analytics now sized by RouteNumRelays instead of the full fixed array; removed dead FirstUpdate flag (never set) and dead RelayManager.TotalCounters (never written); gcloud shutdown-detection ticker 100ms -> 1s (each tick shells out to gcloud, was a busy loop); ip.To4() nil guards for IPv6 clients in dev locate/ISP paths; client relay selection no longer truncates relay lat/long to whole degrees; RelayUpdateResponsePacket.GetMaxSize budgets 19 bytes per address for IPv6; RouteExists compared the full fixed-size relay array instead of routeNumRelays entries

+++ modules/common/client_relays.go
@@ -32,8 +32,8 @@ func GetClientRelays(maxClientRelays int, distanceThreshold int, latencyThreshol
 	for i, relayId := range relayIds {
 		clientRelayData[i].Id = relayId
 		clientRelayData[i].Address = relayAddresses[i]
-		clientRelayData[i].Latitude = float64(int64(relayLatitudes[i]))
-		clientRelayData[i].Longitude = float64(int64(relayLongitudes[i]))
+		clientRelayData[i].Latitude = float64(relayLatitudes[i])
+		clientRelayData[i].Longitude = float64(relayLongitudes[i])
 		clientRelayData[i].Distance = int(core.HaversineDistance(float64(sourceLatitude), float64(sourceLongitude), float64(clientRelayData[i].Latitude), float64(clientRelayData[i].Longitude)))
 	}
